+++ index.html
@@ -85,9 +85,7 @@
             <p class="popup__signature"></p>
         </div>
     </div>
-<script src="./js/validate.js"></script>
-<script src="./js/initial-сards.js"></script>
-<script src="./js/card.js"></script>
-<script src="./js/script.js"></script>
+<script type="module" src="./js/index.js"></script>
+<script type="module" src="./js/Card.js"></script>
 </body>
 </html>
